UPDATE - Characteristic component name
