+++ src/ui/pages/Landing/Landing.tsx
@@ -8,7 +8,7 @@ import { routes } from 'configuration/routes';
 
 import { Button } from 'ui/components/Button';
 import { Logo } from 'ui/components/Logo';
-import { MainPoint } from './components/MainPoint/MainPoint';
+import { Characteristic } from './components/Characteristic';
 import { WhyTwenti } from './components/WhyTwenti';
 
 export const Landing = () => {
@@ -27,12 +27,17 @@ export const Landing = () => {
             comunicarse entre ellas y compartir información.
           </p>
         </Styles.Description>
-        <Styles.MainPoints>
-          <MainPoint description="Conéctate, comparte y comunícate con tus amigos, compañeros de trabajo y familia." icon={social} title="Social" />
-          <MainPoint description="Descubre servicios locales y participa con las marcas que realmente te importan." icon={local} title="Local" />
-          <MainPoint description="Accede a Tuenti desde tu móvil en tiempo real estés donde estés." icon={mobile} title="Móvil" />
-        </Styles.MainPoints>
+        <Styles.List>
+          <Characteristic
+            description="Conéctate, comparte y comunícate con tus amigos, compañeros de trabajo y familia."
+            icon={social}
+            title="Social"
+          />
+          <Characteristic description="Descubre servicios locales y participa con las marcas que realmente te importan." icon={local} title="Local" />
+          <Characteristic description="Accede a Tuenti desde tu móvil en tiempo real estés donde estés." icon={mobile} title="Móvil" />
+        </Styles.List>
       </Styles.Container>
+
       <Styles.ActionButtons>
         <WhyTwenti />
         <Button onClick={redirect}>Pide tú invitación</Button>
