+++ src/ui/pages/Landing/components/Characteristic/@types/Characteristic.types.ts
@@ -1,0 +1,5 @@
+export interface CharacteristicProps {
+  description: string;
+  icon: string;
+  title: string;
+}
