+++ src/ui/pages/Landing/components/Characteristic/index.ts
@@ -1,0 +1,1 @@
+export { Characteristic } from './Characteristic';
