--- src/ui/pages/Landing/components/MainPoint/index.ts
@@ -1,1 +1,0 @@
-export { MainPoint } from './MainPoint';
